+++ Tablut/src/it/unibo/ai/didattica/competition/tablut/droptablut/interfaces/IHeuristic.java
@@ -3,5 +3,5 @@ package it.unibo.ai.didattica.competition.tablut.droptablut.interfaces;
 import it.unibo.ai.didattica.competition.tablut.domain.State;
 
 public interface IHeuristic {
-    double heuristic(State state);
+    float heuristic(State state);
 }
